// src/server/shared/lib/pncp/models/AtaRegistroPreco.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Compra } from './Compra';
export type AtaRegistroPreco = {
  id?: number;
  compra?: Compra;
  numeroAtaRegistroPreco: string;
  anoAta: number;
  dataPublicacaoPncp?: string;
  dataAssinatura?: string;
  cancelado?: boolean;
  dataInclusao?: string;
  dataAtualizacao?: string;
  dataAtualizacaoGlobal?: string;
  dataVigenciaInicio?: string;
  dataVigenciaFim?: string;
  sequencialAta?: number;
  dataCancelamento?: string;
  excluido?: boolean;
  atributoControle?: number;
  numeroControlePNCP?: string;
};