// src/server/shared/lib/pncp/models/RecuperarDadosDocumentoAtaDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type RecuperarDadosDocumentoAtaDTO = {
  url?: string;
  dataPublicacaoPncp?: string;
  titulo?: string;
  sequencialDocumento?: number;
  tipoDocumentoNome?: string;
  tipoDocumentoId?: number;
};